requirements.txt
selenium~=4.9.1
pytest~=7.3.1
webdriver-manager~=3.8.6
allure-pytest
requests

Faker ~= 19.1.0